backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
ebooklib==0.18
beautifulsoup4==4.12.2
Pillow==10.0.0    